The final iteration perhaps. Changed the logic of the bit, and just made it more "consistent".
[Vertical Traversal]
You have these "grid weights". These are like 'width-coordinates' that
starts from left to right, in which are the center of a box — that has a
length of which was divided from the box feed: <box-feed-width> / 50.

These weights determine what will be the next box to be selected
horizontally during vertical traversal. The code would determine which
one of the boxes on the destination row is nearest to a certain grid
weight. It checks the distance of the boxes to the specified grid weight:
<left-side> - <grid-weight>; <right-side> - <grid-weight>; <center-side>
- <grid-weight>.

The results get sorted by ascending then selects the first element of
the three arrays to get the lowest value.

You would see that there's a condition after finding the side having the
lowest value — comparing the founded side if less than 12. This is just
a way to avoid certain occurences that the destination box having its
lowest-value-side being so near at the grid weight. This is why the
<row_sample__c> exists. To compensate that thought.

[Horizontal Traversal]

Just get the next x position then just run that said column to the
weights — you subtract the respective column to the grid weights. Then,
it is sorted by ascending; get the lowest value by selecting the first
element. And then that's it, you have your new grid-weight nearest to
the new horizontally travesed box.